back/requirements.txt
fastapi==0.115.4
uvicorn==0.32.1
pydantic==2.9.0
SQLalchemy==2.0.34
PyMySQL==1.1.1
python-dotenv==1.0.1
cryptography==44.0.0
requests==2.32.3